// src/app/shared/models/permissions.model.ts
export class Permissions {
    admin?: boolean;
    crear_rol?: boolean;
    eliminar_rol?: boolean;
    actualizar_rol?: boolean;
    actualizar_permiso?: boolean;
    eliminar_permiso?: boolean;
    crear_permiso?: boolean;
    eliminar_personal_data?: boolean;
    actualizar_personal_data?: boolean;
    buscar_todos_employees?: boolean;
    crear_employees?: boolean;
    actualizar_employees?: boolean;
    eliminar_employees?: boolean;
    manager?: boolean;
    buscar_todos_rol?: boolean;
    buscar_todos_permiso?: boolean;
    buscar_todos_personal_data?: boolean;
    buscar_todos_card_position?: boolean;
    crear_card_position?: boolean;
    actualizar_card_position?: boolean;
    eliminar_card_position?: boolean;
    buscar_todos_competences?: boolean;
    buscar_todos_fichajes?: boolean;
    crear_competences?: boolean;
    actualizar_competences?: boolean;
    eliminar_competences?: boolean;
    buscar_todos_curriculums?: boolean;
    crear_curriculums?: boolean;
    actualizar_curriculums?: boolean;
    eliminar_curriculums?: boolean;
    buscar_todos_position?: boolean;
    crear_position?: boolean;
    actualizar_position?: boolean;
    eliminar_position?: boolean;
    añadir_campos_extra_personal_data?: boolean;
    buscar_todos_companies?: boolean;
    crear_companies?: boolean;
    actualizar_companies?: boolean;
    eliminar_companies?: boolean;
    buscar_todos_positionHistory?: boolean;
    crear_positionHistory?: boolean;
    actualizar_positionHistory?: boolean;
    eliminar_positionHistory?: boolean;
    upsert_masivo_employees?: boolean;
    show_search_employee?: boolean;
    buscar_todos_unity?: boolean;
    crear_unity?: boolean;
    actualizar_unity?: boolean;
    eliminar_unity?: boolean;
    assess_all?: boolean;
}